backend/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.23.2
torch==2.2.0
torchvision==0.18.1
pillow==10.1.0